requirements.txt
openai>=1.0.0
anthropic>=0.40.0
google-generativeai>=0.8.0
groq>=0.9.0
flask>=3.0.0
python-dotenv>=1.0.0
